+++ lib/unistr/u16-uctomb.c
@@ -28,7 +28,7 @@
 #if !HAVE_INLINE
 
 int
-u16_uctomb (uint16_t *s, ucs4_t uc, int n)
+u16_uctomb (uint16_t *s, ucs4_t uc, ptrdiff_t n)
 {
   if (uc < 0xd800)
     {
